+++ src/lib/mrvGL/mrvTimelineViewportEvents.cpp
@@ -231,10 +231,11 @@ namespace mrv
                         }
                         else
                         {
+                            p.isScrubbing = true;
                             if (Fl::event_alt())
                             {
                                 float multiplier = p.ui->uiPrefs->uiPrefsAltScrubbingSensitivity->value();
-                                scrub(3.0);
+                                scrub(multiplier);
                             }
                             else
                             {
@@ -283,10 +284,10 @@ namespace mrv
                             const int X = Fl::event_x() * pixels_per_unit();
                             const float scale = p.ui->uiPrefs->uiPrefsScrubbingSensitivity->value() * 20;
                             float dx = (X - p.mousePress.x) / scale;
-                            std::cerr << dx << std::endl;
                             
                             if (std::abs(dx) >= 1.0F)
                             {
+                                p.isScrubbing = true;
                                 _scrub(dx);
                                 p.mousePress.x = X;
                             }
@@ -562,7 +563,10 @@ namespace mrv
                         p.actionMode == ActionMode::kRotate)
                     {
                         if (p.player && p.actionMode == ActionMode::kScrub)
+                        {
                             p.player->setPlayback(timeline::Playback::Stop);
+                            p.isScrubbing = true;
+                        }
                         
                         p.lastEvent = FL_PUSH;
                         return;
@@ -1191,9 +1195,11 @@ namespace mrv
                         Fl::event_button() == FL_LEFT_MOUSE && !Fl::event_shift())
                     {
                         p.lastEvent = 0;
+                        p.isScrubbing = false;
                         if (!p.player)
                             return 1;
                         p.player->setPlayback(p.playbackMode);
+                        
                         panel::redrawThumbnails();
                     }
                     else
@@ -1220,6 +1226,7 @@ namespace mrv
                             }
                             else
                             {
+                                p.isScrubbing = false;
                                 p.player->setPlayback(p.playbackMode);
                                 panel::redrawThumbnails();
                             }