+++ src/lib/mrvVk/mrvTimelineViewport.cpp
@@ -98,6 +98,7 @@ namespace mrv
         TimelineViewport::Private::monitorOCIOOptions;
         timeline::Playback TimelineViewport::Private::playbackMode =
             timeline::Playback::Stop;
+        bool TimelineViewport::Private::isScrubbing = false;
         float TimelineViewport::Private::masking = 0.F;
         int64_t TimelineViewport::Private::lastFrame;
         uint64_t TimelineViewport::Private::droppedFrames = 0;
@@ -424,12 +425,15 @@ namespace mrv
             const double fuzzFactor = t.rate() * 0.5;
             const double rate = t.rate() + fuzzFactor;
 
-            if (elapsedTime >= rate)
+            if (elapsedTime >= rate && p.isScrubbing)
             {
                 player->setPlayback(timeline::Playback::Stop);
                 return;
             }
 
+            if (!p.isScrubbing)
+                return;
+            
             Fl::repeat_timeout(
                 1.0 / t.rate(),
                 (Fl_Timeout_Handler)stop_playback_while_scrubbing_cb, this);